# Cleanup PromQL functions