@roidelapluie 2020-01-03
My hardware is probably to blame for the bad benchmarks

```
benchmark                                                                                                             old ns/op       new ns/op       delta
BenchmarkRangeQuery/expr=a_one,steps=1-2                                                                              127281          180747          +42.01%
BenchmarkRangeQuery/expr=a_one,steps=10-2                                                                             135031          138969          +2.92%
BenchmarkRangeQuery/expr=a_one,steps=100-2                                                                            162142          148146          -8.63%
BenchmarkRangeQuery/expr=a_one,steps=1000-2                                                                           405768          398351          -1.83%
BenchmarkRangeQuery/expr=a_ten,steps=1-2                                                                              363781          465278          +27.90%
BenchmarkRangeQuery/expr=a_ten,steps=10-2                                                                             425010          455266          +7.12%
BenchmarkRangeQuery/expr=a_ten,steps=100-2                                                                            728212          792614          +8.84%
BenchmarkRangeQuery/expr=a_ten,steps=1000-2                                                                           3906232         3267234         -16.36%
BenchmarkRangeQuery/expr=a_hundred,steps=1-2                                                                          4656610         5177995         +11.20%
BenchmarkRangeQuery/expr=a_hundred,steps=10-2                                                                         4478043         4722226         +5.45%
BenchmarkRangeQuery/expr=a_hundred,steps=100-2                                                                        7414890         8230034         +10.99%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-2                                                                       36464724        32751643        -10.18%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-2                                                                    139841          154448          +10.45%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-2                                                                   185720          219646          +18.27%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-2                                                                  223582          233995          +4.66%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-2                                                                 801038          799470          -0.20%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-2                                                                    630053          608436          -3.43%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-2                                                                   569443          508780          -10.65%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-2                                                                  1189187         1530095         +28.67%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-2                                                                 7272446         6909456         -4.99%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-2                                                                5211185         5046594         -3.16%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-2                                                               5839539         5551608         -4.93%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-2                                                              13348187        12297050        -7.87%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-2                                                             82437544        74267881        -9.91%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-2                                                                6880917         6978191         +1.41%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10000-2                                                                74379373        73055165        -1.78%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10000-2                                                            665063069       723251218       +8.75%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-2                                                  5215619         5260383         +0.86%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-2                                                 5963249         6333818         +6.21%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-2                                                16331796        18120326        +10.95%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-2                                               131176597       126366794       -3.67%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-2                                                  30197582        29578806        -2.05%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-2                                                 41334074        38532964        -6.78%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-2                                                143129245       167415548       +16.97%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-2                                               1260379887      1277013475      +1.32%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-2                                              261136416       250431812       -4.10%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-2                                             378019562       352946092       -6.63%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-2                                            1441661500      1512803400      +4.93%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-2                                           12181368600     12124039688     -0.47%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-2                                                                 3627149         4491232         +23.82%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-2                                                                4473366         4576174         +2.30%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-2                                                               12130271        10991906        -9.38%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-2                                                              77862632        73423592        -5.70%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-2                                                                 24053243        29028341        +20.68%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-2                                                                31805126        33979499        +6.84%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-2                                                               96250389        144418864       +50.04%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-2                                                              704823206       744649219       +5.65%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-2                                                             238281470       244577565       +2.64%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-2                                                            302769006       308211925       +1.80%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-2                                                           1026747150      1128473125      +9.91%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-2                                                          7183432337      7533192150      +4.87%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-2                                                                    3632666         3538311         -2.60%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-2                                                                   4113183         6410798         +55.86%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-2                                                                  8685874         8500389         -2.14%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-2                                                                 54440933        52338971        -3.86%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-2                                                                    24071442        26079227        +8.34%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-2                                                                   32488053        30307622        -6.71%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-2                                                                  86712615        72153256        -16.79%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-2                                                                 533914969       610207275       +14.29%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-2                                                                258723534       241298667       -6.73%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-2                                                               267766353       312168884       +16.58%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-2                                                              765674075       736689731       -3.79%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-2                                                             5147099938      5355902612      +4.06%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1-2                                                        3686943         3438867         -6.73%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=10-2                                                       3880864         3875915         -0.13%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=100-2                                                      6232735         6005041         -3.65%
BenchmarkRangeQuery/expr=absent_over_time(a_one[1d]),steps=1000-2                                                     28419504        27839635        -2.04%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1-2                                                        29187102        24169856        -17.19%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=10-2                                                       26155340        26121644        -0.13%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=100-2                                                      48631933        48077808        -1.14%
BenchmarkRangeQuery/expr=absent_over_time(a_ten[1d]),steps=1000-2                                                     251847966       306808962       +21.82%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1-2                                                    237657620       222283850       -6.47%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=10-2                                                   280543300       272723273       -2.79%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=100-2                                                  450619100       477858487       +6.04%
BenchmarkRangeQuery/expr=absent_over_time(a_hundred[1d]),steps=1000-2                                                 2729580262      2871860338      +5.21%
BenchmarkRangeQuery/expr=-a_one,steps=1-2                                                                             134947          139287          +3.22%
BenchmarkRangeQuery/expr=-a_one,steps=10-2                                                                            131624          130662          -0.73%
BenchmarkRangeQuery/expr=-a_one,steps=100-2                                                                           167610          159305          -4.95%
BenchmarkRangeQuery/expr=-a_one,steps=1000-2                                                                          383421          425309          +10.92%
BenchmarkRangeQuery/expr=-a_ten,steps=1-2                                                                             501047          537581          +7.29%
BenchmarkRangeQuery/expr=-a_ten,steps=10-2                                                                            471175          441141          -6.37%
BenchmarkRangeQuery/expr=-a_ten,steps=100-2                                                                           798100          788538          -1.20%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-2                                                                          3264361         3277954         +0.42%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-2                                                                         4456468         5136619         +15.26%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-2                                                                        4978353         5513975         +10.76%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-2                                                                       11463336        8129264         -29.08%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-2                                                                      46118464        32565140        -29.39%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-2                                                                      217462          216490          -0.45%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-2                                                                     271780          228481          -15.93%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-2                                                                    567214          545470          -3.83%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-2                                                                   3412793         3681028         +7.86%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-2                                                                      1014233         902112          -11.05%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-2                                                                     1244149         1176485         -5.44%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-2                                                                    3951481         3823218         -3.25%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-2                                                                   28525436        30047696        +5.34%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-2                                                              10063259        9730144         -3.31%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-2                                                             11347842        13289643        +17.11%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-2                                                            39477357        40976901        +3.80%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-2                                                           370968513       354034579       -4.56%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-2                                                                  32775869        35607321        +8.64%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-2                                                                  290929134       275752069       -5.22%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-2                                                          3473654163      3795475463      +9.26%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-2                                                     406483          420861          +3.54%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-2                                                    513923          390395          -24.04%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-2                                                   585368          619089          +5.76%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-2                                                  2137098         2246922         +5.14%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-2                                                     1396377         1004655         -28.05%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-2                                                    1183849         1127081         -4.80%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-2                                                   3192775         2789277         -12.64%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-2                                                  15074871        15194239        +0.79%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-2                                             7727096         9500074         +22.94%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-2                                            8619527         10153431        +17.80%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-2                                           23266086        27019709        +16.13%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-2                                          167704163       168964598       +0.75%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-2                                                      396619          413179          +4.18%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-2                                                     477391          489863          +2.61%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-2                                                    798481          797571          -0.11%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-2                                                   3274294         2699546         -17.55%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-2                                                      1051127         1483991         +41.18%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-2                                                     1476399         1192396         -19.24%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-2                                                    3396744         3298006         -2.91%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-2                                                   21595959        22455748        +3.98%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-2                                              7321117         7163648         -2.15%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-2                                             11674717        8807603         -24.56%
…